[SC-72] refactor client socket
