+++ src/client.js
@@ -1,5 +1,4 @@
 import io from "socket.io-client";
 
 var socket = io();
-console.log(socket);
 export { socket };
